Check for file existance if _isSophonPreloadCompleted being set

+++ CollapseLauncher/Classes/InstallManagement/BaseClass/InstallManagerBase.cs
@@ -126,8 +126,11 @@ namespace CollapseLauncher.InstallManager.Base
                         return;
                     }
 
-                    fileInfo.IsReadOnly = false;
-                    fileInfo.Delete();
+                    if (fileInfo.Exists)
+                    {
+                        fileInfo.IsReadOnly = false;
+                        fileInfo.Delete();
+                    }
                 }
                 catch (Exception ex)
                 {
